manual fix after merge.
Former-commit-id: d97896c115726eaa623fdf4b68db5d4269807dd9

+++ src/UI/Capi/WB.UI.Capi/SynchronizationActivity.cs
@@ -408,7 +408,7 @@ namespace WB.UI.Capi
                                         string.Format(
                                             Resources.GetString(Resource.String.PleaseCheckURLInSettingsFormat),
                                             settingsManager.GetSyncAddressPoint(), GetNetworkDescription(),
-                                            GetNetworkStatus(restException.StatusCode)));
+                                            GetNetworkStatus((int)restException.StatusCode)));
                                 }
                                 else
                                 {
